add Week1 Step3
