+++ CodeStarterCamp_Week1/step2.swift
@@ -1,0 +1,21 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+import Foundation
+
+func drawIcecream(rowIcecream : Int) {
+    for count in 1...rowIcecream{
+        print("***********")}
+}
+func drawBar(rowBar : Int) {
+    for count in 1...rowBar{
+        print("    | |")}
+}
+
+drawIcecream(rowIcecream: 8)
+
+drawBar(rowBar: 4)
